+++ js/data.js
@@ -1,0 +1,24 @@
+// data.js
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const NAMES = ['Артём', 'Мария', 'Дмитрий', 'Анна', 'Сергей', 'Елена', 'Иван', 'Ольга', 'Алексей', 'Татьяна'];
+
+export const DESCRIPTIONS = [
+  'Отличный момент!',
+  'Закат на море',
+  'Мой любимый питомец',
+  'Семейное фото',
+  'Путешествие по горам',
+  'Городские огни',
+  'Природа в её лучшем проявлении',
+  'Архитектурный шедевр',
+  'Уличное искусство',
+  'Фото на память'
+];
