+++ js/utils.js
@@ -1,0 +1,11 @@
+export function getRandomInteger(min, max) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export function generateUniqueIds(count) {
+  const ids = new Set();
+  while (ids.size < count) {
+    ids.add(getRandomInteger(1, 1000));
+  }
+  return [...ids];
+}
